+++ client/src/lib/stores/sandbox/squadStore.ts
@@ -1,0 +1,5 @@
+import { writable } from 'svelte/store';
+
+const initSquads: Record<number, Array<Unit>> = {};
+
+export const squads = writable(initSquads);
