+++ client/src/types.d.ts
@@ -1,0 +1,5 @@
+type Unit = {
+  cost: number;
+  id: number;
+  selected: boolean;
+};
